+++ packages/dara-components/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   Fixed an issue where `Switch` would not be aligned by default with other components
+-   Fixed an issue where `Select` did not accept `items` containing a `DerivedVariable` with a list of strings.
 ## 1.0.0-a.1
 
 -   Initial release